Extract identity-changing inventory mutations
